MiningCatalyst/backend/requirements.txt
alembic==1.9.4
anyio==3.6.2
asn1crypto==1.5.1
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
fastapi==0.92.0
filelock==3.9.0
greenlet==2.0.2
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
oscrypto==1.3.0
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
sniffio==1.3.0
snowflake-connector-python==3.0.0
snowflake-sqlalchemy==1.4.6
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
